airflow/requirements.txt
apache-airflow==2.7.3
boto3==1.34.0
psycopg2-binary==2.9.9
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
slack-sdk==3.26.2